Update for layers/renders and major GUI changes
Now handles Layers and Renders together or separately.
The GUI can dock with KiCad. A great place to dock is right under the Layer Manager.
There are tool tips on all buttons, and includes the layer list tool tip on the View Set button.
Lots of cruft of experiments in this release. It doesn't affect the use of the tool, but this will be cleaned up in a future release.